rename page-size.js to page-style.js
