+++ src/js/viewmodels/settings-panel.js
@@ -20,7 +20,7 @@
 
 import ko from "knockout";
 import ViewerOptions from "../models/viewer-options";
-import PageStyle from "../models/page-size";
+import PageStyle from "../models/page-style";
 import PageViewMode from "../models/page-view-mode";
 import {Keys} from "../utils/key-util";
